django-emails/requirements.txt
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.6
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.10
django-crispy-forms==2.0
django-otp==1.2.2
django-pandas==0.6.6
django-redis==5.3.0
django-rest-passwordreset==1.3.0
django-tempus-dominus==5.1.2.17
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
firebase-admin==6.2.0
Flask==2.3.2
future==0.18.2
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.25.1
oauthlib==3.2.0
olefile==0.46
openshot-qt==3.1.1
pandas==2.0.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
proto-plus==1.22.3
protobuf==4.23.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.4.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.27
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.6.0
reportlab==3.6.8
requests==2.25.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==1.4.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.2.4
systemd-python==234
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==2.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
